Moved commits from Branch 5 to this Branch
GitHub is bugging hard so I just made this same branch so I can make a PR to main again since I tried making a PR with branch 5 but it wasn't recognizing the changes
