+++ frontend/public/RizzUp Icon with Background.png
